requirements.txt
flask
werkzeug
tensorflow-cpu 
gunicorn
numpy